Add jekyll-redirect-from plugin configuration
jekyll-redirect-from plugin allows to define in jekyll's format matter header links which must redirect to or from the current page. This plugin is supported by github pages so no additional configuration must be done apart from declaring it in the _config.yml

As it is an external plugin, it is a different dependency that must be suffied. Hence, Gemfile is added so the dependency management is easier. Before generating the documentation, invoke bundle install command so the dependencies are updated. The Gemfile does not define any dependency itself, but uses github-pages gem which sets up and maintains a local Jekyll environment in sync with GitHub Pages